requirements.txt
gunicorn
flask
psycopg2-binary
imdbpy
fnmatch